requirements.txt
flask
gunicorn
pandas
openpyxl
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
pyarrow